Update README.mkd
I thought we discussed this and decide to separate all words with dashes. Maybe it's just me but findinstore isn't particularly readable to my eye. Of course this is open to discussion and input from the entire group and if the consensus is that we should keep things as is then that's fine. But I wanted to open this issue up just to make sure we are all on the same page.