Removed unneeded docker build/tag/push scripts.
These scripts shouldn't be used anymore with the Vagrant/Docker setup in our local dev env (they were meant to be used with boot2docker).
Also, no one should need to push an image built from its local machine. It is Go responsibility of doing this and GO is not using these scripts for doing that.