Logback configuration for tests.
test/resources/logger.xml overrides conf/logger.xml to log by default at DEBUG (instead of INFO) level and append to STDOUT (instead of FILE) when running tests (both in the IDE and via sbt).

This meant as an easier way to see logs during tests execution rather than having to tail the log file and/or temporarily change conf/logger.xml to set the level to DEBUG and to append to stdout.